+++ CHANGELOG.md
@@ -12,6 +12,7 @@
 - Change frame sleep logic to get fps closer to target fps
 - Auto-repeat bug, which forced to disable xlib auto-repeat during the game execution
 - Encode the first frame when starting encode before the game is launched
+- Encode non-draw frames at game startup as blank frames
 
 ## [1.3.0] - 2018-07-30
 ### Added